fix transfer-encoding for large DynamoDB downloads; update base Dockerfile (#482)
