+++ bin/Dockerfile.base
@@ -27,9 +27,9 @@ RUN set -x && apk add --no-cache openjdk8="$JAVA_ALPINE_VERSION" \
 
 # Install Maven - taken from official repo:
 # https://github.com/carlossg/docker-maven/blob/master/jdk-8/Dockerfile)
-ARG MAVEN_VERSION=3.5.0
+ARG MAVEN_VERSION=3.5.2
 ARG USER_HOME_DIR="/root"
-ARG SHA=beb91419245395bd69a4a6edad5ca3ec1a8b64e41457672dc687c173a495f034
+ARG SHA=707b1f6e390a65bde4af4cdaf2a24d45fc19a6ded00fff02e91626e3e42ceaff
 ARG BASE_URL=https://apache.osuosl.org/maven/maven-3/${MAVEN_VERSION}/binaries
 RUN mkdir -p /usr/share/maven /usr/share/maven/ref \
   && curl -fsSL -o /tmp/apache-maven.tar.gz ${BASE_URL}/apache-maven-$MAVEN_VERSION-bin.tar.gz \
@@ -52,12 +52,10 @@ RUN mkdir -p /opt/code/localstack/localstack/infra && \
         https://artifacts.elastic.co/downloads/elasticsearch/elasticsearch-5.3.0.zip && \
     (cd localstack/infra/ && unzip -q /tmp/localstack.es.zip && \
         mv elasticsearch* elasticsearch && rm /tmp/localstack.es.zip) && \
-
     mkdir -p /opt/code/localstack/localstack/infra/dynamodb && \
     wget -O /tmp/localstack.ddb.zip \
         https://s3-us-west-2.amazonaws.com/dynamodb-local/dynamodb_local_latest.zip && \
     (cd localstack/infra/dynamodb && unzip -q /tmp/localstack.ddb.zip && rm /tmp/localstack.ddb.zip) && \
-
     (pip install --upgrade pip) && \
     (test `which virtualenv` || \
         pip install virtualenv || \
