+++ tests/integration/test_dynamodb.py
@@ -6,6 +6,7 @@ from localstack.utils.aws import aws_stack
 from localstack.utils.common import json_safe
 
 TEST_DDB_TABLE_NAME = 'test-ddb-table-1'
+TEST_DDB_TABLE_NAME_2 = 'test-ddb-table-2'
 PARTITION_KEY = 'id'
 
 
@@ -32,3 +33,23 @@ class DynamoDBIntegrationTest (unittest.TestCase):
             item1 = json_safe(item)
             item2 = json_safe(items[item_id])
             assert item1 == item2
+
+    def test_large_data_download(self):
+        dynamodb = aws_stack.connect_to_resource('dynamodb')
+        dynamodb_client = aws_stack.connect_to_service('dynamodb')
+
+        testutil.create_dynamodb_table(TEST_DDB_TABLE_NAME_2, partition_key=PARTITION_KEY)
+        table = dynamodb.Table(TEST_DDB_TABLE_NAME_2)
+
+        # Create a large amount of items
+        num_items = 20
+        for i in range(0, num_items):
+            item = {PARTITION_KEY: 'id%s' % i, 'data1': 'foobar123 ' * 1000}
+            table.put_item(Item=item)
+
+        # Retrieve the items. The data will be transmitted to the client with chunked transfer encoding
+        result = table.scan(TableName=TEST_DDB_TABLE_NAME_2)
+        assert len(result['Items']) == num_items
+
+        # Clean up
+        dynamodb_client.delete_table(TableName=TEST_DDB_TABLE_NAME_2)
